refactor(frontend): remove low-fanout barrel exports
Migrate all callers of the remaining convenience index.ts barrels to
direct file imports and delete the barrels:

- components/{devices,activities,permissions,roles,import,active,auth}
- components/ui/database
- test/fixtures

vi.mock paths in the affected page tests now target the concrete
component files. The broad public entrypoints lib/swr and server/auth
are intentionally left untouched.

Closes #1776
