+++ frontend/src/app/[tenant]/(protected)/active-supervisions/page.part5.test.tsx
@@ -135,7 +135,7 @@ vi.mock("~/components/sse/SSEErrorBoundary", () => ({
 }));
 
 // Mock UnclaimedRooms
-vi.mock("~/components/active", () => ({
+vi.mock("~/components/active/unclaimed-rooms", () => ({
   UnclaimedRooms: () => <div data-testid="unclaimed-rooms" />,
 }));
