+++ frontend/src/app/[tenant]/(protected)/database/devices/page.tsx
@@ -21,7 +21,7 @@ import { getDbOperationMessage } from "@/lib/use-notification";
 import { createCrudService } from "@/lib/database/service-factory";
 import { devicesConfig } from "@/components/database/configs/devices.config";
 import { getDeviceTypeDisplayName, type Device } from "@/lib/iot-helpers";
-import { DevicesMasterDetail } from "@/components/devices";
+import { DevicesMasterDetail } from "@/components/devices/devices-master-detail";
 import { DatabaseFormModal } from "~/components/ui/database/database-form-modal";
 import { ConfirmationModal } from "~/components/ui/modal";
 import { useToast } from "~/contexts/ToastContext";
