+++ frontend/src/app/[tenant]/(protected)/database/roles/page.tsx
@@ -16,9 +16,9 @@ import { createCrudService } from "@/lib/database/service-factory";
 import { rolesConfig } from "@/components/database/configs/roles.config";
 import type { Role } from "@/lib/auth-helpers";
 import { getRoleDisplayName } from "@/lib/auth-helpers";
-import { RolesMasterDetail } from "@/components/roles";
+import { RolesMasterDetail } from "@/components/roles/roles-master-detail";
 import { DatabaseFormModal } from "~/components/ui/database/database-form-modal";
-import { RolePermissionManagementModal } from "@/components/auth";
+import { RolePermissionManagementModal } from "@/components/auth/role-permission-management-modal";
 import { ConfirmationModal } from "~/components/ui/modal";
 import { useToast } from "~/contexts/ToastContext";
 import { useIsMobile } from "~/components/ui/hooks/useIsMobile";
