+++ frontend/src/app/[tenant]/(protected)/database/students/import/page.test.tsx
@@ -70,7 +70,7 @@ vi.mock("~/components/ui/alert", () => ({
 }));
 
 // Mock import components
-vi.mock("~/components/import", () => ({
+vi.mock("~/components/import/upload-section", () => ({
   UploadSection: ({
     onFileSelect,
     isDragging,
@@ -107,6 +107,9 @@ vi.mock("~/components/import", () => ({
       </button>
     </div>
   ),
+}));
+
+vi.mock("~/components/import/stats-cards", () => ({
   StatsCards: ({
     total,
     newCount,
@@ -125,6 +128,9 @@ vi.mock("~/components/import", () => ({
       <span data-testid="stat-errors">{errors}</span>
     </div>
   ),
+}));
+
+vi.mock("~/components/import/student-row-card", () => ({
   StudentRowCard: ({
     student,
   }: {
