--- frontend/src/components/active/index.ts
@@ -1,1 +1,0 @@
-export { UnclaimedRooms } from "./unclaimed-rooms";
