--- frontend/src/components/activities/index.ts
@@ -1,1 +1,0 @@
-export { ActivitiesMasterDetail } from "./activities-master-detail";
