--- frontend/src/components/auth/index.ts
@@ -1,2 +1,0 @@
-export * from "./role-permission-management-modal";
-export * from "./invitation-accept-form";
