--- frontend/src/components/devices/index.ts
@@ -1,1 +1,0 @@
-export { DevicesMasterDetail } from "./devices-master-detail";
