--- frontend/src/components/import/index.ts
@@ -1,3 +1,0 @@
-export { UploadSection } from "./upload-section";
-export { StatsCards } from "./stats-cards";
-export { StudentRowCard } from "./student-row-card";
