--- frontend/src/components/permissions/index.ts
@@ -1,2 +1,0 @@
-export { PermissionEditModal } from "./permission-edit-modal";
-export { PermissionsMasterDetail } from "./permissions-master-detail";
