--- frontend/src/components/roles/index.ts
@@ -1,1 +1,0 @@
-export { RolesMasterDetail } from "./roles-master-detail";
