--- frontend/src/components/ui/database/index.ts
@@ -1,4 +1,0 @@
-// Database component library exports
-export * from "./database-form";
-export * from "./database-form-modal";
-export { GroupSelect } from "./database-select";
