--- frontend/src/test/fixtures/index.ts
@@ -1,13 +1,0 @@
-export { buildBackendStudent, buildStudent } from "./students";
-
-export { buildBackendGroup } from "./groups";
-
-export { buildBackendRoom } from "./rooms";
-
-export { buildBackendActivity, buildBackendCategory } from "./activities";
-
-export {
-  buildBackendActiveSession,
-  buildBackendVisit,
-  buildBackendSupervisor,
-} from "./sessions";
